Up to 9.3
Added control flow logic of Ifs! Learned the reason why the
conditions for if statements in C like languages are surrounded
in parentheses. It was a way for Dennis Ritchie to delineate
the end of the condition from the start of the if branch while
maintaining balanced brackets. There's no need for the opening '('
since we already have the 'if' token.